refactor(boot/boot.asm): delete 'jmp $+2' ins
The jmps in between are for synchronizing CPU and PIC, so the CPU won't send data the PIC cannot receive yet.
This is comparable to wait states when writing to memory: when the CPU is faster than the memory/memory arbiter, it needs to wait some time before accessing memory the next time.
Maybe can read data from some special IO port to judgment PIC status, fix later.
References:
1. https://stackoverflow.com/questions/34561275/setting-up-interrupts-in-protected-mode-x86